third_party/blink/web_tests/external/wpt/webrtc/RTCPeerConnection-ondatachannel-expected.txt

This is a testharness.js-based test.
PASS Data channel event should fire when new data channel is announced to the remote peer
PASS Should be able to send data in a datachannel event handler
FAIL Open event should not be raised when closing the channel in the datachannel event assert_unreached: Open event should not fire Reached unreachable code
PASS Open event should be raised when closing the channel in the datachannel event after enqueuing a task
FAIL Open event should not be raised when sending and immediately closing the channel in the datachannel event assert_unreached: Open event should not fire Reached unreachable code
FAIL In-band negotiated channel created on remote peer should match the same configuration as local peer assert_equals: expected (string) "high" but got (undefined) undefined
FAIL In-band negotiated channel created on remote peer should match the same (default) configuration as local peer assert_equals: expected (string) "low" but got (undefined) undefined
PASS Negotiated channel should not fire datachannel event on remote peer
Harness: the test ran to completion.
